refactor(error): 境界の HTTP 写像を汎用ヘルパー＋対応表に集約する
集約ごとに近い実装の throwXHttpError を 3 つ持っていたが、ドメインエラー → HTTP
ステータスの対応表を渡すだけの汎用 throwHttpError に一本化する。対応表は集約ごとに
AUTH_ERROR_STATUS / ARTICLE_ERROR_STATUS / ACCOUNT_ERROR_STATUS として持つ。

- 各エラーモジュール(auth-error.ts 等)と what コメントを廃止し、throw-http-error.ts へ集約
- テストの captureThrow を test/helper へ切り出し共有化
