+++ application/apps/web/src/server/passkey/handler/registration-verify.ts
@@ -3,7 +3,7 @@ import { authClientConfig, PasskeyClient } from '@trend-diary/authentication'
 import { z } from 'zod'
 import CONTEXT_KEY from '@/middleware/context'
 import zodValidator, { type ZodValidatedContext } from '@/middleware/zod-validator'
-import throwHttpError from '@/server/error/auth-error'
+import throwHttpError, { AUTH_ERROR_STATUS } from '@/server/error/throw-http-error'
 
 // 真正性はSupabaseが検証するため中身の妥当性検証はプロバイダに委ね、ここは登録 ceremony 結果を素通しする
 export const passkeyRegistrationVerifyInputSchema = z.object({
@@ -29,7 +29,7 @@ export default async function passkeyRegistrationVerify(
     challengeId: valid.challengeId,
     credential: valid.credential,
   })
-  if (result.isErr()) throwHttpError(result.error)
+  if (result.isErr()) throwHttpError(result.error, AUTH_ERROR_STATUS)
 
   logger.info('passkey registration success', { passkeyId: result.value.id })
