+++ application/apps/web/src/server/passkey/passkey-action.ts
@@ -2,7 +2,7 @@ import { type AuthError, authClientConfig, PasskeyClient } from '@trend-diary/au
 import type { Context } from 'hono'
 import type { Result } from 'neverthrow'
 import type { Env } from '@/env'
-import throwHttpError from '@/server/error/auth-error'
+import throwHttpError, { AUTH_ERROR_STATUS } from '@/server/error/throw-http-error'
 
 export type PasskeyActionContext = Context<Env>
 
@@ -13,7 +13,7 @@ export function createPasskeyActionHandler<TOutput, TResponse>(config: {
   return async (c: PasskeyActionContext) => {
     const passkeyClient = new PasskeyClient(authClientConfig(c))
     const result = await config.execute(passkeyClient)
-    if (result.isErr()) throwHttpError(result.error)
+    if (result.isErr()) throwHttpError(result.error, AUTH_ERROR_STATUS)
 
     return config.respond(c, result.value)
   }
