+++ application/apps/web/src/server/sessions/handler/create.ts
@@ -4,8 +4,10 @@ import { authInputSchema, createAccountUseCase } from '@trend-diary/domain/accou
 import CONTEXT_KEY from '@/middleware/context'
 import zodValidator, { type ZodValidatedContext } from '@/middleware/zod-validator'
 import { assertCaptchaVerified } from '@/server/captcha'
-import throwAccountHttpError from '@/server/error/account-error'
-import throwAuthHttpError from '@/server/error/auth-error'
+import throwHttpError, {
+  ACCOUNT_ERROR_STATUS,
+  AUTH_ERROR_STATUS,
+} from '@/server/error/throw-http-error'
 
 export const authInputValidator = zodValidator('json', authInputSchema)
 
@@ -20,14 +22,14 @@ export default async function createSession(c: ZodValidatedContext<[typeof authI
 
   const authClient = new PasswordAuthClient(authClientConfig(c))
   const loginResult = await authClient.signIn({ email: valid.email, password: valid.password })
-  if (loginResult.isErr()) throwAuthHttpError(loginResult.error)
+  if (loginResult.isErr()) throwHttpError(loginResult.error, AUTH_ERROR_STATUS)
 
   const user = loginResult.value
 
   const rdb = getRdbClient(c.env.DB)
   const accountUseCase = createAccountUseCase(rdb)
   const result = await accountUseCase.resolveActiveUser(user.id)
-  if (result.isErr()) throwAccountHttpError(result.error)
+  if (result.isErr()) throwHttpError(result.error, ACCOUNT_ERROR_STATUS)
 
   logger.info('session created', { activeUserId: result.value.activeUserId })
