+++ application/apps/web/src/server/sessions/handler/destroy.ts
@@ -1,14 +1,14 @@
 import { authClientConfig, PasswordAuthClient } from '@trend-diary/authentication'
 import type { Context } from 'hono'
 import CONTEXT_KEY from '@/middleware/context'
-import throwHttpError from '@/server/error/auth-error'
+import throwHttpError, { AUTH_ERROR_STATUS } from '@/server/error/throw-http-error'
 
 export default async function destroySession(c: Context) {
   const logger = c.get(CONTEXT_KEY.APP_LOG)
 
   const authClient = new PasswordAuthClient(authClientConfig(c))
   const result = await authClient.signOut()
-  if (result.isErr()) throwHttpError(result.error)
+  if (result.isErr()) throwHttpError(result.error, AUTH_ERROR_STATUS)
 
   logger.info('session destroyed')
